chore: [#186749997] refactored isRemoteSellerRemoteWorker to be used in more places
