+++ shared/src/index.ts
@@ -4,8 +4,8 @@ export * from "./businessUser";
 export * from "./countries";
 export * from "./dateHelpers";
 export * from "./defaultConstants";
+export * from "./domain-logic/businessPersonaHelpers";
 export * from "./domain-logic/createBusinessId";
-export * from "./domain-logic/determineForeignBusinessType";
 export * from "./domain-logic/fetchMunicipalityById";
 export * from "./domain-logic/getCurrentBusiness";
 export * from "./domain-logic/hasCompletedBusinessStructure";
