+++ web/src/components/Header.tsx
@@ -3,11 +3,11 @@ import { AuthContext } from "@/contexts/authContext";
 import { IsAuthenticated } from "@/lib/auth/AuthContext";
 import { useConfig } from "@/lib/data-hooks/useConfig";
 import { useUserData } from "@/lib/data-hooks/useUserData";
-import { isStartingBusiness } from "@/lib/domain-logic/businessPersonaHelpers";
 import { ROUTES } from "@/lib/domain-logic/routes";
 import analytics from "@/lib/utils/analytics";
 import { templateEval } from "@/lib/utils/helpers";
 import { getCurrentDateInNewJersey } from "@businessnjgovnavigator/shared/";
+import { isStartingBusiness } from "@businessnjgovnavigator/shared/domain-logic/businessPersonaHelpers";
 import { LookupIndustryById } from "@businessnjgovnavigator/shared/industry";
 import { LookupLegalStructureById } from "@businessnjgovnavigator/shared/legalStructure";
 import { useRouter } from "next/router";
